+++ lib/dradis/plugins/nexpose.rb
@@ -7,7 +7,8 @@ end
 
 require 'dradis/plugins/nexpose/engine'
 require 'dradis/plugins/nexpose/field_processor'
-require 'dradis/plugins/nexpose/importer'
+require 'dradis/plugins/nexpose/full/importer'
 require 'dradis/plugins/nexpose/mapping'
+require 'dradis/plugins/nexpose/simple/importer'
 require 'dradis/plugins/nexpose/version'
 require 'dradis/plugins/nexpose/xml_formatter'