+++ lib/tasks/thorfile.rb
@@ -1,10 +1,27 @@
 class NexposeTasks < Thor
   include Rails.application.config.dradis.thor_helper_module
 
-  namespace "dradis:plugins:nexpose"
+  namespace 'dradis:plugins:nexpose:upload'
 
-  desc "upload FILE", "upload NeXpose results"
-  def upload(file_path)
+  desc 'full FILE', 'upload NeXpose full results'
+  def full(file_path)
+    detect_and_set_project_scope
+
+    importer = Dradis::Plugins::Nexpose::Full::Importer.new(task_options)
+    importer.import(file: file_path)
+  end
+
+  desc 'simple FILE', 'upload NeXpose simple results'
+  def simple(file_path)
+    detect_and_set_project_scope
+
+    importer = Dradis::Plugins::Nexpose::Simple::Importer.new(task_options)
+    importer.import(file: file_path)
+  end
+
+  private
+
+  def process_upload(importer, file_path)
     require 'config/environment'
 
     unless File.exists?(file_path)
@@ -12,10 +29,6 @@ class NexposeTasks < Thor
       exit -1
     end
 
-    detect_and_set_project_scope
-
-    importer = Dradis::Plugins::Nexpose::Importer.new(task_options)
     importer.import(file: file_path)
   end
-
 end