+++ spec/nexpose_upload_spec.rb
@@ -6,79 +6,52 @@ describe 'Nexpose upload plugin' do
     @fixtures_dir = File.expand_path('../fixtures/files/', __FILE__)
   end
 
-  describe 'importer' do
-    before(:each) do
-      # Stub template service
-      templates_dir = File.expand_path('../../templates', __FILE__)
-      expect_any_instance_of(Dradis::Plugins::TemplateService)
-        .to receive(:default_templates_dir).and_return(templates_dir)
-
-      # Init services
-      plugin = Dradis::Plugins::Nexpose
-
-      @content_service = Dradis::Plugins::ContentService::Base.new(
-        logger: Logger.new(STDOUT),
-        plugin: plugin
+  describe 'Importer: Full with fingerprints elements' do
+    def apply_mapping(doc, field)
+      ms = Dradis::Plugins::MappingService.new(
+        integration: Dradis::Plugins::Nexpose
       )
+      mapping_fields = [OpenStruct.new(
+        destination_field: 'Fingerprints',
+        content: "{{ nexpose[#{field}] }}"
+      )]
 
-      @importer = plugin::Importer.new(
-        content_service: @content_service,
+      @result = ms.apply_mapping(
+        data: doc.at_xpath('//nodes/node'),
+        mapping_fields: mapping_fields,
+        source: 'full_node'
       )
+    end
 
-      # Stub dradis-plugins methods
-      #
-      # They return their argument hashes as objects mimicking
-      # Nodes, Issues, etc
-      allow(@content_service).to receive(:create_node) do |args|
-        OpenStruct.new(args)
+    describe 'with fingerprints > OS elements' do
+      it 'uses the os product value' do
+        doc = Nokogiri::XML(File.read(@fixtures_dir + '/full.xml'))
+        apply_mapping(doc, 'node.fingerprints')
+
+        expect(@result).to include('IOS')
       end
-      allow(@content_service).to receive(:create_note) do |args|
-        OpenStruct.new(args)
-      end
-      allow(@content_service).to receive(:create_issue) do |args|
-        OpenStruct.new(args)
-      end
-      allow(@content_service).to receive(:create_evidence) do |args|
-        OpenStruct.new(args)
+
+      it 'defaults to n/a if there is no os product value' do
+        doc = Nokogiri::XML(File.read(@fixtures_dir + '/full_with_duplicate_node.xml'))
+        apply_mapping(doc, 'node.fingerprints')
+
+        expect(@result).to include('n/a')
       end
     end
 
-    describe 'Importer: Simple' do
-      it 'creates nodes, issues, notes and an evidences as needed' do
+    describe 'with software > fingerprint elements' do
+      it 'uses the product value given software/fingerprints' do
+        doc = Nokogiri::XML(File.read(@fixtures_dir + '/full.xml'))
+        apply_mapping(doc, 'node.software')
 
-        expect(@content_service).to receive(:create_node).with(hash_including label: '1.1.1.1', type: :host).once
+        expect(@result).to include('JRE')
+      end
 
-        expect(@content_service).to receive(:create_note) do |args|
-          expect(args[:text]).to include('Host Description : Linux 2.6.9-89.ELsmp')
-          expect(args[:text]).to include('Scanner Fingerprint certainty : 0.80')
-          expect(args[:node].label).to eq('1.1.1.1')
-        end.once
+      it 'defaults to n/a if there is no os product value' do
+        doc = Nokogiri::XML(File.read(@fixtures_dir + '/full_with_duplicate_node.xml'))
+        apply_mapping(doc, 'node.software')
 
-        expect(@content_service).to receive(:create_node) do |args|
-          expect(args[:label]).to eq('Generic Findings')
-          expect(args[:parent].label).to eq('1.1.1.1')
-          OpenStruct.new(args)
-        end.once
-
-        expect(@content_service).to receive(:create_node) do |args|
-          expect(args[:label]).to eq('udp-000')
-          expect(args[:parent].label).to eq('1.1.1.1')
-          OpenStruct.new(args)
-        end.once
-
-        expect(@content_service).to receive(:create_note) do |args|
-          expect(args[:text]).to include("#[Id]#\nntpd-crypto")
-          expect(args[:text]).to include("#[host]#\n1.1.1.1")
-          expect(args[:node].label).to eq('udp-000')
-        end.once
-
-        expect(@content_service).to receive(:create_note) do |args|
-          expect(args[:text]).to include("#[Id]#\nntp-clock-radio")
-          expect(args[:text]).to include("#[host]#\n1.1.1.1")
-          expect(args[:node].label).to eq('udp-000')
-        end.once
-
-        @importer.import(file: @fixtures_dir + '/simple.xml')
+        expect(@result).to include('n/a')
       end
     end
 
@@ -170,7 +143,7 @@ describe 'Nexpose upload plugin' do
           expect(args[:content]).to include("#[Content]#\nThe following NTP variables")
           OpenStruct.new(args)
         end.once
-  
+
         expect(@content_service).to receive(:create_evidence) do |args|
           expect(args[:content]).to include("#[Content]#\nVulnerable URL:")
           OpenStruct.new(args)