+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-v4.12.0 (Month 2024)
+v4.16.0 (Month 2025)
   - Format UnorderedList/OrderedList to textile
+
+v4.15.0 (December 2024)
+  - No changes
+
+v4.14.0 (October 2024)
+- Separate general importer into Full & Simple importers
+
+v4.13.0 (July 2024)
+  - No changes
 
 v4.12.0 (May 2024)
  - Migrate integration to use Mappings Manager
  - Update Dradis links in README
+ - Use 'n/a' as OS node property if there is no product value
 
 v4.11.0 (January 2024)
   - Add port/protocol to evidences